LayoutTests/fast/js/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A6.3_T3-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Parse error
S7.9_A6.3_T3

PASS Expected parsing failure

TEST COMPLETE
